Bubbles - Spheres and Wireframes
I was originally playing around with the concept of having circles with
just lines outside the logo, and just filled with the color inside the
logo. To achieve the look I wanted, I really needed to draw the outside
ones first and then the opaque ones on top. I decided to try making it
3D and throwing a sphere in instead of an ellipse, and I think it was a
happy accident :)