Angel Camoflage
A quick variation on the previous Voronoi sketch. I had the values map
inverse those outside. Interestingly, when it comes close to middle
grey, it looks like those shapes are on top of the letters. This is
because it freely crosses over the bounds of the letters and background
as one shape (see Josef Albers, The Interaction of Color).